gen/traffic-generator/norm/requirements.txt
scipy
numpy
pyyaml